+++ libraries/AS5600/CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.5.1] - 2023-12-31
+- fix #51, add **increaseOffset(float degrees)**
+- update keywords.txt
+- update readme.md (several cleanups)
+
+
 ## [0.5.0] - 2023-12-07
 - refactor API, begin()
 - update readme.md